[CREATE] Date Ext

+++ bamboo-iOS/Sources/Common/Models/Algorithem/Algorithem.swift
@@ -13,7 +13,7 @@ struct Algorithem : ModelType{
     var title : String
     var content : String
     var tag : String
-    var createdAt : Int64
+    var createdAt : Int
     var reason : String?
     var isDone = false
 }
